# models/User.go
// user.go

package models

import "time"

type User struct {
	ID            uint       `gorm:"primaryKey"`
	Username      string     `gorm:"column:user_name"`
	Password      string     `gorm:"column:pass"`
	Realname      string     `gorm:"column:real_name"`
	Phone         string     `gorm:"column:phone"`
	Status        int        `gorm:"column:status"`
	Remark        string     `gorm:"column:remark"`
	LastLoginTime *time.Time `gorm:"column:last_login_time"`
	LastLoginIP   string     `gorm:"column:last_login_ip"`
	LoginTimes    int        `gorm:"column:login_times"`
	CreatedAt     time.Time  `gorm:"column:created_at" json:",omitempty"`
	UpdatedAt     time.Time  `gorm:"column:updated_at"`
	Avatar        string     `gorm:"column:avatar"`
}

type ResponseUser struct {
	ID       uint   `gorm:"primaryKey"`
	Username string `gorm:"column:user_name"`
	Phone    string `gorm:"column:phone"`
	Avatar   string `gorm:"column:avatar"`
}

func (User) TableName() string {
	return "ga_users"
}